fix safelen for non 64-bit targets

+++ src/tests.rs
@@ -118,8 +118,8 @@ fn doctest3() {
 
     a.append(&mut b);
 
-    assert_eq!(a.len(), 5usize);
-    assert_eq!(b.len(), 0usize);
+    assert_eq!(a.len() as usize, 5usize);
+    assert_eq!(b.len() as usize, 0usize);
 
     assert!(a.contains(1));
     assert!(a.contains(2));
@@ -403,7 +403,7 @@ fn understand_bitand_assign() {
 #[test]
 fn iters() {
     let range_set_blaze = RangeSetBlaze::from_iter([1..=6, 8..=9, 11..=15]);
-    assert!(range_set_blaze.len() == 13usize);
+    assert!(range_set_blaze.len() as usize == 13usize);
     for i in range_set_blaze.iter() {
         println!("{i}");
     }
@@ -879,8 +879,8 @@ fn lib_coverage_0() {
 
     a.append(&mut b);
 
-    assert_eq!(a.len(), 5usize);
-    assert_eq!(b.len(), 0usize);
+    assert_eq!(a.len() as usize, 5usize);
+    assert_eq!(b.len() as usize, 0usize);
 
     assert!(a.contains(1));
     assert!(a.contains(2));
@@ -946,7 +946,7 @@ fn lib_coverage_0() {
     assert_eq!(set.ranges_insert(2..=5), true);
     assert_eq!(set.ranges_insert(5..=6), true);
     assert_eq!(set.ranges_insert(3..=4), false);
-    assert_eq!(set.len(), 5usize);
+    assert_eq!(set.len() as usize, 5usize);
     let mut set = RangeSetBlaze::from_iter([1, 2, 3]);
     assert_eq!(set.take(2), Some(2));
     assert_eq!(set.take(2), None);
@@ -963,11 +963,11 @@ fn lib_coverage_0() {
 
     let mut a = RangeSetBlaze::from_iter([1..=3]);
     a.extend(std::iter::once(4));
-    assert_eq!(a.len(), 4usize);
+    assert_eq!(a.len() as usize, 4usize);
 
     let mut a = RangeSetBlaze::from_iter([1..=3]);
     a.extend(4..=5);
-    assert_eq!(a.len(), 5usize);
+    assert_eq!(a.len() as usize, 5usize);
 
     let mut set = RangeSetBlaze::new();
 
@@ -1001,7 +1001,7 @@ fn lib_coverage_0() {
 
     let mut a = RangeSetBlaze::from_iter([1..=3]);
     a.extend([1..=3]);
-    assert_eq!(a.len(), 3usize);
+    assert_eq!(a.len() as usize, 3usize);
 
     let a = RangeSetBlaze::from_iter([1..=3]);
     let b = <RangeSetBlaze<i32> as Clone>::clone(&a);
